Don't use bootstrap chroot on mock 1.4
Mock 1.4 adds a feature to create a bootstrap chroot of the target
system and then use that to create the actual requested chroot
using the target system's version of the package manager. Since it
doesn't support microdnf and we don't have full DNF in the module,
we need to skip the bootstrap.

Signed-off-by: Stephen Gallagher <sgallagh@redhat.com>